tighten rules regarding c++'s type conversion function used in converting container index and issue a variety of diagnostics when such conversions can not be done. Ths should conclude code changes. Few more tests to follow.

+++ test/CodeGenObjCXX/objc-container-subscripting.mm
@@ -21,7 +21,7 @@ struct S {
 int main() {
   NSMutableArray<P> * array;
   S s;
-  id oldObject = array[s];
+  id oldObject = array[(int)s];
 
   NSMutableDictionary<P> *dict;
   dict[(id)s] = oldObject;
